Fyp-part1-main-main/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
shap==0.42.1
matplotlib==3.7.2
seaborn==0.12.2
nltk==3.8.1
spacy==3.6.1
python-dotenv==1.0.0
joblib==1.3.1
imbalanced-learn==0.11.0
scipy==1.11.2
# For spacy model
# python -m spacy download en_core_web_sm